% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{launch}
\alias{launch}
\title{Run contained Shiny app in a web browser}
\usage{
launch()
}
\description{
\code{vinylspotting} provides an interactive tool to explore
Discogs collections. The tool will be launched in a web browser.
}
